Reorganized and alphabetized statues, misc cleanup
